[IMP] partner_statement: add hooks to make logic more inheritable
