+++ partner_statement/report/__init__.py
@@ -2,6 +2,7 @@ from . import report_statement_common
 from . import activity_statement
 from . import detailed_activity_statement
 from . import outstanding_statement
+from . import report_statement_common_xlsx
 from . import activity_statement_xlsx
 from . import detailed_activity_statement_xlsx
 from . import outstanding_statement_xlsx
